fix(web-ui): stop pagination offering a Next link past the end
PageResponse.hasNext compared page + 1 to totalPages in int arithmetic. The
page number is echoed back from the backend, and every paged endpoint
bounds it below at zero with @Min(0) while none bounds it above - so a
request for page 2147483647 came back with that number intact, page + 1
wrapped to a negative, and a negative is below any page count. The view
then rendered a Next control on an empty result set beyond the end of the
data.

Compare as longs so the increment cannot wrap. The link builder already
clamps its target at zero, so the visible symptom was a Next control that
sent the operator back to the first page rather than a broken URL - still
wrong, and confusing on an admin listing.

Covered by tests for the wrapping page, an ordinary page, the last page and
an empty result.

+++ web-ui-service/src/main/java/com/orderprocessing/webui/dto/PageResponse.java
@@ -21,5 +21,11 @@ public record PageResponse<T>(
 
     public int number() { return page; }
     public boolean hasPrevious() { return page > 0; }
-    public boolean hasNext() { return page + 1 < totalPages; }
+
+    /**
+     * Compared as longs on purpose. The page number is echoed back from the backend, which bounds it below at
+     * zero but not above, so a request for page 2147483647 made {@code page + 1} wrap to a negative int - which
+     * is less than any page count, so the view offered a Next link past the end of an empty result.
+     */
+    public boolean hasNext() { return (long) page + 1L < (long) totalPages; }
 }
